Extract CLI and wizard from clip_manager into corridorkey_cli
Separate the interactive wizard and CLI argument parsing from the
pipeline library so clip_manager.py can be imported without side
effects. This enables VFX pipeline integration — Nuke scripts,
Houdini plugins, and batch automation can now do:

    from clip_manager import scan_clips, run_inference, ClipEntry

New file: corridorkey_cli.py
  - interactive_wizard() (moved from clip_manager.py)
  - _configure_environment() (warnings, logging setup)
  - main() entry point with argparse

clip_manager.py is now a pure library (893 lines, down from 1157).
All launcher scripts updated to point to corridorkey_cli.py.
Added 'corridorkey' entry point in pyproject.toml.
