+++ clip_manager.py
@@ -1,10 +1,8 @@
-import argparse
 import glob
 import logging
 import os
 import shutil
 import sys
-import warnings
 
 # Enable OpenEXR support in OpenCV — needed for EXR I/O throughout the pipeline.
 # Must be set before any cv2.imread/imwrite calls on .exr files.
@@ -14,17 +12,6 @@ import cv2
 import numpy as np
 
 logger = logging.getLogger(__name__)
-
-
-def _configure_environment():
-    """Set up logging and warnings for interactive CLI use.
-
-    Called from __main__ only — importing clip_manager as a library
-    does not suppress warnings or configure the root logger.
-    Library callers should configure logging themselves.
-    """
-    warnings.filterwarnings("ignore")
-    logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
 
 
 # Core Paths
@@ -864,236 +851,6 @@ def organize_clips(clips_dir):
             organize_target(full_path)
 
 
-def interactive_wizard(win_path):
-    print("\n" + "=" * 60)
-    print(" CORRIDOR KEY - SMART WIZARD")
-    print("=" * 60)
-
-    # 1. Map Path
-    linux_path = map_path(win_path)
-    print(f"Windows Path: {win_path}")
-    print(f"Linux Path:   {linux_path}")
-
-    if not os.path.exists(linux_path):
-        print("\n[ERROR] Path does not exist on Linux mount!")
-        print(f"Expected: {LINUX_MOUNT_ROOT}")
-        return
-
-    # 2. Analyze
-    # We treat linux_path as the ROOT containing SHOTS
-    # Or is linux_path the SHOT itself?
-    # HEURISTIC: If it contains "Input", it's a shot. If it contains folders, it's a project.
-
-    # Let's assume it's a folder containing CLIPS (Batch Mode)
-    # But if the user drops it IN a shot folder, we should handle that too.
-
-    target_is_shot = False
-    if os.path.exists(os.path.join(linux_path, "Input")) or glob.glob(os.path.join(linux_path, "Input.*")):
-        target_is_shot = True
-
-    work_dirs = []
-    if target_is_shot:
-        work_dirs = [linux_path]
-    else:
-        # Scan subfolders
-        work_dirs = [
-            os.path.join(linux_path, d) for d in os.listdir(linux_path) if os.path.isdir(os.path.join(linux_path, d))
-        ]
-        # Filter out output/hints
-        work_dirs = [
-            d
-            for d in work_dirs
-            if os.path.basename(d) not in ["Output", "AlphaHint", "VideoMamaMaskHint", ".ipynb_checkpoints"]
-        ]
-
-    print(f"\nFound {len(work_dirs)} potential clip folders.")
-
-    # Check for loose videos in root
-    loose_videos = [
-        f for f in os.listdir(linux_path) if is_video_file(f) and os.path.isfile(os.path.join(linux_path, f))
-    ]
-
-    # Check if existing folders need organization
-    dirs_needing_org = []
-    for d in work_dirs:
-        # Check for Input
-        has_input = os.path.exists(os.path.join(d, "Input")) or glob.glob(os.path.join(d, "Input.*"))
-        # Check for hints
-        has_alpha = os.path.exists(os.path.join(d, "AlphaHint"))
-        has_mask = os.path.exists(os.path.join(d, "VideoMamaMaskHint"))
-
-        if not has_input or not has_alpha or not has_mask:
-            dirs_needing_org.append(d)
-
-    if loose_videos or dirs_needing_org:
-        if loose_videos:
-            print(f"Found {len(loose_videos)} loose video files that need organization:")
-            for v in loose_videos:
-                print(f"  - {v}")
-
-        if dirs_needing_org:
-            print(f"Found {len(dirs_needing_org)} folders that might need setup (Hints/Input):")
-            # Limit output if too many
-            if len(dirs_needing_org) < 10:
-                for d in dirs_needing_org:
-                    print(f"  - {os.path.basename(d)}")
-            else:
-                print(f"  - ...and {len(dirs_needing_org)} others.")
-
-        # 3. Organize Loop
-        yn = input("\n[1] Organize Clips & Create Hint Folders? [y/N]: ").strip().lower()
-        if yn == "y":
-            # Organize loose videos first
-            for v in loose_videos:
-                clip_name = os.path.splitext(v)[0]
-                ext = os.path.splitext(v)[1]
-                target_folder = os.path.join(linux_path, clip_name)
-
-                if os.path.exists(target_folder):
-                    logger.warning(f"Skipping loose video '{v}': Target folder '{clip_name}' already exists.")
-                    continue
-
-                try:
-                    os.makedirs(target_folder)
-                    target_file = os.path.join(target_folder, f"Input{ext}")
-                    shutil.move(os.path.join(linux_path, v), target_file)
-                    logger.info(f"Organized: Moved '{v}' to '{clip_name}/Input{ext}'")
-
-                    # Also initialize hints immediately
-                    for hint in ["AlphaHint", "VideoMamaMaskHint"]:
-                        os.makedirs(os.path.join(target_folder, hint), exist_ok=True)
-                except Exception as e:
-                    logger.error(f"Failed to organize video '{v}': {e}")
-
-            # Organize existing folders
-            for d in work_dirs:
-                organize_target(d)
-            print("Organization Complete.")
-
-            # Re-scan in case structure changed
-            # If it was a shot, it's still a shot (unless we messed it up)
-            # If it wasn't a shot, we re-scan subdirs
-            if not target_is_shot:
-                work_dirs = [
-                    os.path.join(linux_path, d)
-                    for d in os.listdir(linux_path)
-                    if os.path.isdir(os.path.join(linux_path, d))
-                ]
-                work_dirs = [
-                    d for d in work_dirs if os.path.basename(d) not in ["Output", "AlphaHint", "VideoMamaMaskHint"]
-                ]
-
-    # 4. Status Check Loop
-    while True:
-        ready = []
-        masked = []
-        raw = []
-
-        for d in work_dirs:
-            entry = ClipEntry(os.path.basename(d), d)
-            try:
-                entry.find_assets()  # This checks Input and AlphaHint
-            except (FileNotFoundError, ValueError, OSError):
-                pass  # Might act up if Input missing
-
-            # Check VideoMamaMaskHint (Strict: videomamamaskhint.ext or VideoMamaMaskHint/)
-            has_mask = False
-            mask_dir = os.path.join(d, "VideoMamaMaskHint")
-
-            # 1. Directory Check
-            if os.path.isdir(mask_dir) and len(os.listdir(mask_dir)) > 0:
-                has_mask = True
-
-            # 2. File Check (Strict Stem Match)
-            if not has_mask:
-                for f in os.listdir(d):
-                    stem, _ = os.path.splitext(f)
-                    if stem.lower() == "videomamamaskhint" and is_video_file(f):
-                        has_mask = True
-                        break
-
-            if entry.alpha_asset:
-                ready.append(entry)
-            elif has_mask:
-                masked.append(entry)
-            else:
-                raw.append(entry)
-
-        print("\n" + "-" * 40)
-        print("STATUS REPORT:")
-        print(f"  READY (AlphaHint found): {len(ready)}")
-        for c in ready:
-            print(f"    - {c.name}")
-
-        print(f"  MASKED (VideoMamaMaskHint found): {len(masked)}")
-        for c in masked:
-            print(f"    - {c.name}")
-
-        print(f"  RAW (Input only):        {len(raw)}")
-        for c in raw:
-            print(f"    - {c.name}")
-        print("-" * 40 + "\n")
-
-        # Combine checks for actions
-        missing_alpha = masked + raw
-
-        print("\nACTIONS:")
-        if missing_alpha:
-            print(f"  [v] Run VideoMaMa (Found {len(masked)} ready with masks)")
-            print(f"  [g] Run GVM (Auto-Matte on {len(raw)} clips without Mask Hint)")
-
-        if ready:
-            print(f"  [i] Run Inference (on {len(ready)} ready clips)")
-
-        print("  [r] Re-Scan Folders")
-        print("  [q] Quit")
-
-        choice = input("\nSelect Action: ").strip().lower()
-
-        if choice == "v":
-            # VideoMaMa
-            print("\n--- VideoMaMa ---")
-            print("Scanning for VideoMamaMaskHints...")
-            # We pass ALL missing alpha clips. run_videomama checks for the actual files.
-            run_videomama(missing_alpha, chunk_size=50)
-            input("VideoMaMa batch complete. Press Enter to Re-Scan...")
-            continue
-
-        elif choice == "g":
-            # GVM
-            print("\n--- GVM Auto-Matte ---")
-            print(f"This will generate alphas for {len(raw)} clips that have NO Mask Hint.")
-
-            yn = input("Proceed with GVM? [y/N]: ").strip().lower()
-            if yn == "y":
-                generate_alphas(raw)
-                input("GVM batch complete. Press Enter to Re-Scan...")
-            continue
-
-        elif choice == "i":
-            # Inference
-            print("\n--- Corridor Key Inference ---")
-            try:
-                run_inference(ready)
-            except (RuntimeError, FileNotFoundError) as e:
-                logger.error(f"Inference failed: {e}")
-            input("Inference batch complete. Press Enter to Re-Scan...")
-            continue
-
-        elif choice == "r":
-            print("\nRe-scanning...")
-            continue
-
-        elif choice == "q":
-            break
-
-        else:
-            print("Invalid selection.")
-            continue
-
-    print("\nWizard Complete. Goodbye!")
-
-
 def scan_clips():
     if not os.path.exists(CLIPS_DIR):
         os.makedirs(CLIPS_DIR, exist_ok=True)
@@ -1134,34 +891,3 @@ def scan_clips():
         print("\nAll clip folders appear valid.\n")
 
     return valid_clips
-
-
-if __name__ == "__main__":
-    _configure_environment()
-
-    parser = argparse.ArgumentParser(description="CorridorKey Clip Manager")
-    parser.add_argument("--action", choices=["generate_alphas", "run_inference", "list", "wizard"], required=True)
-    parser.add_argument("--win_path", help=r"Windows Path (example: V:\...) for Wizard Mode", default=None)
-
-    args = parser.parse_args()
-
-    try:
-        if args.action == "list":
-            scan_clips()
-        elif args.action == "generate_alphas":
-            clips = scan_clips()
-            generate_alphas(clips)
-        elif args.action == "run_inference":
-            clips = scan_clips()
-            run_inference(clips)
-        elif args.action == "wizard":
-            if not args.win_path:
-                print("Error: --win_path required for wizard.")
-            else:
-                interactive_wizard(args.win_path)
-    except KeyboardInterrupt:
-        print("\nInterrupted.")
-        sys.exit(130)
-    except Exception as e:
-        logger.error(str(e))
-        sys.exit(1)
